+++ NSTUQuizMaker/database/migrations/2019_05_11_061042_create_teachers_table.php
@@ -17,7 +17,7 @@ class CreateTeachersTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('designation');
-            $table->integer('department_id')->unsigned();
+            $table->bigInteger('department_id')->unsigned();
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
